backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
llama-index>=0.8.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
transformers>=4.30.0
torch>=2.0.0
pypdf>=3.15.0